Moved handlers into blueprint so they can be expanded on later
