+++ dashboard/handlers/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+handler_bp = Blueprint('handlers', __name__, template_folder='templates')
+
+from . import views
