--- dashboard/templates/200.html
@@ -1,8 +1,0 @@
-<!-- extend base layout -->
-{% extends "base.html" %}
-
-{% block content %}
-  <h1>Happy</h1>
-  <p>The administrator has been notified. Sorry for the inconvenience!</p>
-  <p><a href="{{ url_for('main.index') }}">Back</a></p>
-{% endblock %}
